Fix MSVC narrowing warnings in the v2.0 HEIF/AVIF code paths
The v2.0.0 container work added call sites that narrow size_t to the int
lengths libheif's API takes, and none of them had been compiled at /W3
before. Adds the missing casts so the warning-level change in the previous
commit leaves the core target warning-free:

  heifultrahdr.cpp, avifultrahdr.cpp  heif_context_add_exif_metadata and
                                      heif_context_encode_gain_map_image
  ultrahdr_api.cpp                    heif_check_filetype (x2),
                                      heif_image_handle_get_gain_map_metadata_size,
                                      and the exif_size argument of
                                      parseGainMapMetadata
  jpegr.cpp                           the same parseGainMapMetadata argument
  jpegrutils.cpp                      pow() returns double; altHeadroom is float

Wrapped by hand rather than with clang-format: these files are already 55 to
92 lines away from the repo's own .clang-format, so running it would bury the
change in unrelated reformatting. Every touched line is within the 100 column
limit.

+++ lib/src/jpegr.cpp
@@ -1499,7 +1499,7 @@ uhdr_error_info_t JpegR::decodeJPEGR(uhdr_compressed_image_t* uhdr_compressed_im
         static_cast<uint8_t*>(jpeg_dec_obj_gm.getIsoMetadataPtr()),
         jpeg_dec_obj_gm.getIsoMetadataSize(), static_cast<uint8_t*>(jpeg_dec_obj_gm.getXMPPtr()),
         jpeg_dec_obj_gm.getXMPSize(), static_cast<uint8_t*>(jpeg_dec_obj_sdr.getEXIFPtr()),
-        jpeg_dec_obj_sdr.getEXIFSize(), &uhdr_metadata))
+        static_cast<int>(jpeg_dec_obj_sdr.getEXIFSize()), &uhdr_metadata))
     if (gainmap_metadata != nullptr) {
       std::copy(uhdr_metadata.min_content_boost, uhdr_metadata.min_content_boost + 3,
                 gainmap_metadata->min_content_boost);
